refactor, patientservice now uses storage service
