+++ stack2/frontend/journal-react/src/services/PatientService.js
@@ -1,5 +1,6 @@
 import axios from 'axios';
 import AccountService from "./AccountService";
+import StoreService from "./StoreService";
 
 const Request = axios.create({
 	baseURL: 'http://localhost/stack2/patient',
@@ -10,16 +11,27 @@ const Request = axios.create({
 
 const PatientService = {
 	getPatients() {
-		return new Promise((success, fail) => {
+		return new Promise((success) => {
+			let patients = [];
 			Request.get('', {
 				params: {
 				apiKey: AccountService.getApiKey()
 				}
 			}).then((response) => {
-				success(response.data);
-			}).catch(() => {
-				fail("failed");
-			})
+				patients = response.data;
+			}).finally(() => {
+				StoreService.getPatients().then((stored) => {
+					patients = patients.filter((patient) => {
+						const isDuplicated = stored.some((store) => {
+							return store.id === patient.id;
+						});
+						return !isDuplicated;
+						});
+
+					const all = stored.concat(patients);
+					success(all);
+				})
+			});
 		});
 	},
 	getPatient(patientId) {
